Changing the lhttpc load balancer to use ETS
The current implementation uses a dict and a queue for common
socket operations when load-balancing. Over heavy load, the process
gets to be very slow. Plus, it set itself as a high priority process,
unbalancing the whole VM.

This switches the dict to an ETS table, and the queue to a stack
(list) in order to reduce operations. Moreover, the process will
go back to a normal priority to make sure it doesn't mess up with
the schedulers and timers too much.