Adding control flow on connections refused
When many connections were being refused, the load balancer
would impose no good control flow mecanism on incoming requests.
After a while, demand can overtake the process and grow the
message queue until the VM goes out of memory.

This patch adds a counter of refused connections (happens when
the server is down); if too many connections are refusd in a row
(as many as the possible sockets), some of the requests will be
blocked and will return {error, offline}.

Whenever a successful request is made, the counter is dropped.

The patch also contains a few minute optimizations for record
assignment, gaining minimal amounts of speed.