Attempt at fixing a possible process leak
Yoshihiro Tanaka found that when lhttpc:request closes a worker
due to a timeout, it's possible that it happens after the port
is unlinked in prim_inet:close, but before it is properly closed.

This results in orphaned sockets/ports being left hanging in the
ether.

This fix attempts to wrap lhttpc_sock:close commands around a safe
build that should resolve it.

A potential fix would have been to have the manager monitor the
sockets itself, but this wouldn't have worked if the socket is new
and the manager has never seen it before, hence the current fix.