Disable invalid failing tests.
The connection_timeout and connection_count tests uses interfaces that
don't exist in the lhttpc_lb implementation.

The suspended_manager test seems to assume a single connection manager
instead of a per-host/port/ssl manager.  I'm not sure it makes sense in
the current design.  Why would we want to suspend the manager?