lhttpc_dns: New module for DNS caching.
The Erlang DNS client is inefficient, because it requires a call to a
port program inet_gethost, and thus an OS context switch, for each
lookup.  This change adds a cache on the DNS lookups that respects the
DNS TTL.

    otp/lib/kernel/src/inet_gethost_native.erl

Bugs: the current implementation never evicts old entries; it only
grows.