Write host/port/ssl to process dictionary in lhttpc_lb for debugging.  Version 1.8.0.
We're having a problem with a resource leak in the ssl module in Erlang/OTP 20
that results in the lhttpc_lb process's message queue growing.  With this
commit the lhttpc_lb process stores the host, port, and ssl status to the
lhttpc_lb_host in its process dictionary so that we can fetch this from
another process for monitoring.