lhttpc_dns: New logic for DNS servers that return a subset of all addresses on each request.  Version 1.12.0.
This change is made necessary by a provider that has a pool of IP addresses
for a domain name but only returns one out of the set for each DNS lookup.
This would not be a problem in itself, but it appears that they are also doing
some kind of client-IP-based connection-limiting on their end.  This means
that new connections get limited when we use the same remote IP address for
all of our connections.

Now we detect when the results of the DNS lookup change with each request and
we collect the results of all of the recent responses to reconstruct the pool
locally.