Set the SNI TLS extension for SSL connections.  Version 1.13.0.
Erlang automatically sets the SNI (server name indication) TLS extension when
the Host argument to the ssl:connect call is a hostname, but our DNS cache
converts the hostname to an IP address which prevents that from working.  We
now set the server_name_indication option to the original hostname in the
ssl:connect call.