Return load balancer from lhttpc_lb:checkout call, pass back in lhttpc_lb:checkin.
The lhttpc_lb:checkout call now returns the load balancer pid so that we can
pass it back to the lhttpc_lb:checkin call instead of having to look it up a
second time.  This is also in preparation for potentially having multiple load
balancers for a {Host, Port, Ssl} combination.