Use multiple load balancers for large connection pools.
Closing a socket is sometimes slow, and currently the load balancer process
blocks while the connection is closed.  This is the first attempt at solving
this problem.  It does so by using multiple load balancers processes for each
Host/Port/SSL combination, one of which is selected at random to handle each
request.