Fix socket leak that happens when client dies.
The change introduced in the previous commit causes a socket to be leaked when
the client was killed by the timeout mechanism.  In this case, the client
didn't explicitly close the socket but since the socket's controlling process
is now the load balancer the socket was not closed by the Erlang VM.  This fix
is to have the load balancer keep track of the socket and to have it
explicitly close the socket when the client dies.