Fix potential socket leak that happens when client dies.
We need to return the socket to the load balancer before making it the
controlling process to prevent a race.  If we transfer ownership first and
then are killed before we return the socket (the old order), then Erlang VM
will not close the socket until the load balancer exits.  If we return the
socket first and are killed before we can transfer ownership (the new order),
then the Erlang VM will close the socket and the load balancer will receive
the closed socket, which a future client will have to deal with.