Expire idle sockets on checkin instead of using timers.
Lhttpc has a mechanism to shrink the socket pool back down when it grows.

Previously lhttpc_lb started a timer each time a socket was checked back in
and canceled it when the socket was checked out and, closed the socket if the
timer expired before the socket was reused.  In addition, the idle sockets
were kept on a stack, so that when the size of the pool was grown because of a
temporary surge in traffic, the sockets on the bottom of the stack were
unlikely to be reused once the traffic surge was over and those sockets would
have a chance to have their timer expire.

Now, instead of using timers, lhttpc has a stack of socket checkin
times (implemented using Erlang's queue module).  When a socket is checked in
the current time is pushed on the top of the stack and when a socket is
checked out the most recent time is popped from the top of the stack.  In
addition, whenever a socket is checked in the bottom of the checkin time stack
is checked to determine whether the pool is larger than it needs to be to
handle the currently request rate, and sockets are closed as long as the
oldest checkin time is over the idle threshold.

The new mechanism keeps the idle sockets in a FIFO queue instead of stack so
that the requests are spread over all of the currently open sockets.  This
also means that if the remote end closes the socket while it is checked in
that socket will be passed to a client which will notice that the socket is
closed and open a new one instead of having the half-closed socket hang around
on the bottom of the stack.

This original mechanism did have the advantage that the socket pool would be
shrunk immediatly when the idle time timer expired even if the pool stopped
getting requests.  The new mechanism should have less overhead than the old
one since it doesn't start and cancel all those timers.