Make close_connection_timeout stats recording more efficient.  Version 1.17.1.
Previously the lhttpc_lb did not keep a socket when it was checked out.  This
meant that when the client died (due to a timeout, for example) the
lhttpc_stats module had to search the stats table based on the client pid
instead of using the key.  Now that the lhttpc_lb process keeps the sockets
that it checks out the lhttpc_stats process can now use the socket instead of
the pid when recording timeouts.